v1.23 Don't delete SADMIN server directories if SADM_HOST_TYPE='D' in sadmin.cfg

+++ bin/sadm_client_housekeeping.sh
@@ -29,9 +29,10 @@
 # 2018_09_28    v1.17 Code Optimize and Cleanup
 # 2018_10_27    v1.18 Remove old dir. not use anymore (if exist)
 # 2018_11_02    v1.19 Code Maint. & Comment
-#@2018_11_23    v1.20 An error is signal when the 'sadmin' account is lock, preventing cron to run.
-#@2018_11_24    v1.21 Check SADM_USER status, give more precise explanation & corrective cmd if lock.
-#@2018_12_15    v1.22 Fix Error Message on MacOS trying to find SADMIN User in /etc/passwd.
+# 2018_11_23    v1.20 An error is signal when the 'sadmin' account is lock, preventing cron to run.
+# 2018_11_24    v1.21 Check SADM_USER status, give more precise explanation & corrective cmd if lock.
+# 2018_12_15    v1.22 Fix Error Message on MacOS trying to find SADMIN User in /etc/passwd.
+# 2018_12_18    v1.23 Don't delete SADMIN server directories if SADM_HOST_TYPE='D' in sadmin.cfg
 #
 # --------------------------------------------------------------------------------------------------
 #
@@ -55,7 +56,7 @@ trap 'sadm_stop 0; exit 0' 2                                            # INTERC
     fi
 
     # CHANGE THESE VARIABLES TO YOUR NEEDS - They influence execution of SADMIN standard library.
-    export SADM_VER='1.22'                              # Current Script Version
+    export SADM_VER='1.23'                              # Current Script Version
     export SADM_LOG_TYPE="B"                            # Writelog goes to [S]creen [L]ogFile [B]oth
     export SADM_LOG_APPEND="N"                          # Append Existing Log or Create New One
     export SADM_LOG_HEADER="Y"                          # Show/Generate Script Header
@@ -232,14 +233,14 @@ dir_housekeeping()
 
     # Remove Old Ver.1 Dir. Not use anymore
     if [ -d "${SADM_DAT_DIR}/sar" ]
-        then sadm_writelog "Directory ${SADM_DAT_DIR}/sar should exist anymore."
+        then sadm_writelog "Directory ${SADM_DAT_DIR}/sar should not exist anymore."
              sadm_writelog "I am deleting it now."
              rm -fr ${SADM_DAT_DIR}/sar
     fi
 
     # Remove Old Ver.1 Dir. Not use anymore
     if [ -d "${SADM_BASE_DIR}/jac" ]
-       then sadm_writelog "Directory ${SADM_BASE_DIR}/jac should exist anymore."
+       then sadm_writelog "Directory ${SADM_BASE_DIR}/jac should not exist anymore."
             sadm_writelog "I am deleting it now."
             rm -fr ${SADM_BASE_DIR}/jac
     fi
@@ -353,19 +354,19 @@ dir_housekeeping()
             sadm_writelog "Remove useless Directories on client (if any) ..."
             #
             # Remove Database Backup Directory on SADMIN Client if it exist
-            if [ -d "$SADM_DBB_DIR" ]
+            if [ -d "$SADM_DBB_DIR" ] && [ "$SADM_HOST_TYPE = "C" ]
                 then sadm_writelog "Directory $SADM_DBB_DIR should exist only on SADMIN server."
                      sadm_writelog "I am deleting it now."
                      rm -fr $SADM_DBB_DIR
             fi
             # Remove Network Scan Directory on SADMIN Client if it exist
-            if [ -d "$SADM_NET_DIR" ]
+            if [ -d "$SADM_NET_DIR" ] && [ "$SADM_HOST_TYPE = "C" ]
                 then sadm_writelog "Directory $SADM_NET_DIR should exist only on SADMIN server."
                      sadm_writelog "I am deleting it now."
                      rm -fr $SADM_NET_DIR
             fi
             # Remove Web Site Directory on SADMIN Client if it exist
-            if [ -d "$SADM_WWW_DIR" ]
+            if [ -d "$SADM_WWW_DIR" ] && [ "$SADM_HOST_TYPE = "C" ]
                 then sadm_writelog "Directory $SADM_WWW_DIR should exist only on SADMIN server."
                      sadm_writelog "I am deleting it now."
                      rm -fr $SADM_WWW_DIR
